+++ src/Header/index.jsx
@@ -2,8 +2,8 @@
  * Created by xy on 15/4/13.
  */
 const assign = require('object-assign');
-// const Const = require('uxcore-const');
 const Tooltip = require('uxcore-tooltip');
+const Icon = require('uxcore-icon');
 const classnames = require('classnames');
 const React = require('react');
 const addEventListener = require('rc-util/lib/Dom/addEventListener');
@@ -118,7 +118,7 @@ class Header extends React.Component {
         </div>}
         placement="top"
       >
-        <i className="kuma-icon kuma-icon-information" />
+        <Icon name="tishi-full" />
       </Tooltip>
     );
   }