+++ HISTORY.md
@@ -2,6 +2,11 @@
 
 ---
 
+## 5.0.0
+
+* `CHANGED` update `uxcore-date-cell-field` to ^0.4.0
+* `CHANGED` update `uxcore-pagiantion` to ~0.5.0
+
 ## 4.0.0
 
 * `CHANGED` update `uxcore-pagiantion` to ~0.4.0